Crud revised

+++ app/Controllers/Http/CustomersController/index.ts
@@ -6,35 +6,40 @@ class CustomersController {
 
   public async create({ request }: HttpContextContract) {
     const data: CustomerData = request.only(this.customerProperties)
-    const { id } = await Customer.create({ ...data })
+    const { id } = await Customer.create(data)
     return { ...data, id }
   }
 
-  public async index() {
-    const customers = await Customer.all()
+  public async index({ request }: HttpContextContract) {
+    const page = request.input('page') || 1
+    const perPage = 10
+    const customers = await Customer.query().paginate(page, perPage)
     return customers
   }
 
-  public async show({ params }: HttpContextContract) {
+  public async show({ response, params }: HttpContextContract) {
     const { id } = params
     const customer = await Customer.find(id)
+    if (!customer) return response.status(404)
     return customer
   }
 
-  public async update({ request, params }: HttpContextContract) {
+  public async update({ request, response, params }: HttpContextContract) {
     const { id } = params
     const data: CustomerData = request.only(this.customerProperties)
-    const customer = await Customer.findOrFail(id)
+    const customer = await Customer.find(id)
+    if (!customer) return response.status(404)
     customer.merge({ ...data })
     await customer.save()
-    return customer
+    return response.status(204)
   }
 
-  public async delete({ params }: HttpContextContract) {
+  public async destroy({ response, params }: HttpContextContract) {
     const { id } = params
-    const customer = await Customer.findOrFail(id)
+    const customer = await Customer.find(id)
+    if (!customer) return response.status(404)
     await customer.delete()
-    return customer
+    return response.status(204)
   }
 }
